update timer unit

+++ nodeSystem.h
@@ -88,7 +88,7 @@ int nodeSystemSave(char* const path);
 int nodeSystemLoad(char* const path);
 void nodeSystemTimerRun();
 void nodeSystemTimerStop();
-void nodeSystemTimerSet(long period);
+void nodeSystemTimerSet(double period);
 void nodeSystemTimerGet();
 char** nodeSystemGetConst(char* const constNode,char* const constPipe,int* retCode);
 char** nodeSystemGetNodeNameList(int* counts);
@@ -104,5 +104,5 @@ int nodeSystemRead(int pipeID,void* buffer);
 int nodeSystemWrite(int pipeID,void* const buffer);
 int nodeSystemAddPipe(char* const pipeName,NODE_PIPE_TYPE type,NODE_DATA_UNIT unit,uint16_t arrayLength,const void* buff);
 int nodeSystemWait();
-long nodeSystemGetPeriod();
+double nodeSystemGetPeriod();
 #endif
